Handle oauth token exchange without refresh token

+++ src/foam/core/session/cron/ExpireSessionsCron.java
@@ -45,7 +45,7 @@ public class ExpireSessionsCron implements ContextAgent {
 
       var oauthSession = (OAuthCredential) oAuthCredentialDAO.find(EQ(OAuthCredential.SESSION_ID, session.getId()));
       if ( session.getLastUsed().getTime() + session.getTtl() <= now.getTime()
-          || ( oauthSession != null && oauthSession.getExpiresAt().getTime() <= now.getTime() ) ) {
+          || ( oauthSession != null && oauthSession.getExpiresAt() != null && oauthSession.getExpiresAt().getTime() <= now.getTime() ) ) {
         logger.debug("Destroyed expired session : " + (String) session.getId());
         localSessionDAO.remove(session);
       } else if ( oauthSession != null && session.getCreated().getTime() + oauthSessionTtl <= now.getTime() ) {
